+++ flang-rt/lib/runtime/io-api.cpp
@@ -33,23 +33,6 @@
 namespace Fortran::runtime::io {
 RT_EXT_API_GROUP_BEGIN
 
-RT_API_ATTRS const char *InquiryKeywordHashDecode(
-    char *buffer, std::size_t n, InquiryKeywordHash hash) {
-  if (n < 1) {
-    return nullptr;
-  }
-  char *p{buffer + n};
-  *--p = '\0';
-  while (hash > 1) {
-    if (p < buffer) {
-      return nullptr;
-    }
-    *--p = 'A' + (hash % 26);
-    hash /= 26;
-  }
-  return hash == 1 ? p : nullptr;
-}
-
 template <Direction DIR>
 RT_API_ATTRS Cookie BeginInternalArrayListIO(const Descriptor &descriptor,
     void ** /*scratchArea*/, std::size_t /*scratchBytes*/,
