+++ flang-rt/unittests/Runtime/Namelist.cpp
@@ -10,7 +10,7 @@
 #include "CrashHandlerFixture.h"
 #include "tools.h"
 #include "flang-rt/runtime/descriptor.h"
-#include "flang/Runtime/io-api-consts.h"
+#include "flang/Runtime/io-api.h"
 #include <algorithm>
 #include <cinttypes>
 #include <complex>
