+++ flang-rt/unittests/Runtime/RuntimeCrashTest.cpp
@@ -13,7 +13,7 @@
 #include "CrashHandlerFixture.h"
 #include "tools.h"
 #include "flang-rt/runtime/terminator.h"
-#include "flang/Runtime/io-api-consts.h"
+#include "flang/Runtime/io-api.h"
 #include "flang/Runtime/transformational.h"
 #include <gtest/gtest.h>
